LayoutTests/inspector/dom-debugger/dom-breakpoint-subtree-modified-remove-expected.txt

CONSOLE MESSAGE: BREAKPOINT ACTION LOG 1
CONSOLE MESSAGE: BREAKPOINT ACTION LOG 2
CONSOLE MESSAGE: BREAKPOINT ACTION LOG 3
CONSOLE MESSAGE: BREAKPOINT ACTION LOG 4
Tests for DOM breakpoints.


== Running test suite: DOMBreakpoint.SubtreeModified
-- Running test case: DOMBreakpoint.SubtreeModified.Remove.BreakpointEnabled
Adding "subtree-modified:1,HTML,1,BODY,1,DIV,0,DIV" DOM Breakpoint...
PAUSED:
PASS: Pause reason should be DOM.
PASS: Pause type should be 'subtree-modified'.
PASS: Pause nodeId should be expected value.
PASS: Should have 'insertion'.
CALL STACK:
0: [F] testSubtreeModified
1: [P] Global Code
-- Running test teardown.

-- Running test case: DOMBreakpoint.SubtreeModified.Remove.BreakpointDisabled
Adding "subtree-modified:1,HTML,1,BODY,1,DIV,0,DIV" DOM Breakpoint...
Disabling breakpoint...
Triggering breakpoint...
PASS: Should not pause for disabled breakpoint.
-- Running test teardown.

-- Running test case: DOMBreakpoint.SubtreeModified.Remove.DebuggerDisabled
Adding "subtree-modified:1,HTML,1,BODY,1,DIV,0,DIV" DOM Breakpoint...
Disabling all breakpoints...
Triggering breakpoint...
Enabling all breakpoints...
PASS: Should not pause when all breakpoints disabled.
-- Running test teardown.

-- Running test case: DOMBreakpoint.SubtreeModified.Remove.RemoveBreakpoint
Adding "subtree-modified:1,HTML,1,BODY,1,DIV,0,DIV" DOM Breakpoint...
Removing breakpoint...
Triggering breakpoint...
PASS: Should not pause for removed breakpoint.
-- Running test teardown.

-- Running test case: DOMBreakpoint.SubtreeModified.Remove.Options.Condition
Adding "subtree-modified:1,HTML,1,BODY,1,DIV,0,DIV" DOM Breakpoint...

Setting condition to 'false'...

Triggering breakpoint...
PASS: Should not pause.

Triggering breakpoint...
PASS: Should not pause.

Setting condition to 'true'...

Triggering breakpoint...
PASS: Should pause.

Triggering breakpoint...
PASS: Should pause.

-- Running test case: DOMBreakpoint.SubtreeModified.Remove.Options.IgnoreCount
Adding "subtree-modified:1,HTML,1,BODY,1,DIV,0,DIV" DOM Breakpoint...

Setting ignoreCount to '2'...

Triggering breakpoint...
PASS: Should not pause.

Triggering breakpoint...
PASS: Should not pause.

Triggering breakpoint...
PASS: Should pause.

Triggering breakpoint...
PASS: Should pause.

-- Running test case: DOMBreakpoint.SubtreeModified.Remove.Options.Action.Log
Adding "subtree-modified:1,HTML,1,BODY,1,DIV,0,DIV" DOM Breakpoint...

Adding log action...

Triggering breakpoint...
PASS: Should execute breakpoint action.
PASS: Should pause.

Editing log action...

Triggering breakpoint...
PASS: Should execute breakpoint action.
PASS: Should pause.

Editing log action...
Enabling auto-continue...

Triggering breakpoint...
PASS: Should execute breakpoint action.
PASS: Should not pause.

Editing log action...

Triggering breakpoint...
PASS: Should execute breakpoint action.
PASS: Should not pause.

-- Running test case: DOMBreakpoint.SubtreeModified.Remove.Options.Actions.Evaluate
Adding "subtree-modified:1,HTML,1,BODY,1,DIV,0,DIV" DOM Breakpoint...

Adding evaluate action...

Triggering breakpoint...
PASS: Should execute breakpoint action.
PASS: Should pause.

Editing evaluate action...

Triggering breakpoint...
PASS: Should execute breakpoint action.
PASS: Should pause.

Editing evaluate action...
Enabling auto-continue...

Triggering breakpoint...
PASS: Should execute breakpoint action.
PASS: Should not pause.

Editing evaluate action...

Triggering breakpoint...
PASS: Should execute breakpoint action.
PASS: Should not pause.
